feat(crypto): add FORKID sighash support for altcoins
Add support for Bitcoin Cash/Ecash, Bitcoin Gold and other FORKID-based
coins by implementing BIP143-style sighash algorithm with fork ID
support. This allows:

1. Handling transactions with SIGHASH_FORKID flag (0x40)
2. Computing correct FORKID sighashes with (fork_id << 8 | sighash_type)
3. Signing PSBTs with BCH/BTG forkid support via new sign_forkid method

The implementation preserves raw sighash types in Signature.sighash_type
field while providing methods to extract the standard base type.
PSBT deserialization now accepts any sighash byte for better compatibility.

Tests include:
- BCH test vectors for FORKID sighash computation
- BitGo cross-chain PSBT fixtures for Bitcoin and altcoins

Issue: BTC-2656

+++ bitcoin/src/blockdata/witness.rs
@@ -794,7 +794,7 @@ mod test {
             hex!("304402207c800d698f4b0298c5aac830b822f011bb02df41eb114ade9a6702f364d5e39c0220366900d2a60cab903e77ef7dd415d46509b1f78ac78906e3296f495aa1b1b541");
         let signature = secp256k1::ecdsa::Signature::from_der(&sig_bytes).unwrap();
         let mut witness = Witness::default();
-        let signature = crate::ecdsa::Signature { signature, sighash_type: EcdsaSighashType::All };
+        let signature = crate::ecdsa::Signature { signature, sighash_type: EcdsaSighashType::All as u32 };
         witness.push_ecdsa_signature(&signature);
         let expected_witness = vec![hex!(
             "304402207c800d698f4b0298c5aac830b822f011bb02df41eb114ade9a6702f364d5e39c0220366900d2a60cab903e77ef7dd415d46509b1f78ac78906e3296f495aa1b1b54101")
